2.0. audio: Audio pipeline is now in 32 bit.      decklink_consumer: Added support for 32 bit audio. git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@1321 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/mixer/mixer.cpp
@@ -96,7 +96,7 @@ struct mixer::implementation : boost::noncopyable
 	std::unordered_map<int, tweened_transform<core::frame_transform>> transforms_;	
 	std::unordered_map<int, blend_mode::type> blend_modes_;
 
-	std::queue<std::pair<boost::unique_future<safe_ptr<host_buffer>>, std::vector<int16_t>>> buffer_;
+	std::queue<std::pair<boost::unique_future<safe_ptr<host_buffer>>, std::vector<int32_t>>> buffer_;
 	
 	const size_t buffer_size_;
