+++ modules/ffmpeg/producer/audio/audio_decoder.h
@@ -43,7 +43,7 @@ public:
 	
 	void push(const std::shared_ptr<AVPacket>& packet);
 	bool ready() const;
-	std::vector<std::shared_ptr<std::vector<int16_t>>> poll();
+	std::vector<std::shared_ptr<std::vector<int32_t>>> poll();
 
 	int64_t nb_frames() const;
